Fix the last tool-tree sign-compare, and size the -Werror extension
C_SyvUpPacHexFileView.cpp had the same shape cleared from core earlier: an
int32_t loop counter compared against .size(). Cast the size to match.

With that gone, a full clean eight-tool build produces exactly two warnings and
neither is ours -- both are -Woverloaded-virtual in the Scanner.cpp that flex
generates inside the vendored Vector_DBC submodule. Every line we own across all
eight tools is warning-free.

So extending -Werror there is not a cleanup project, but it is also not a
one-liner, and docs/TODO.md now records why. The tools set warnings through
CMAKE_CXX_FLAGS and add_compile_options, which are directory-scoped and propagate
into add_subdirectory targets including the submodule. Appending -Werror would
turn those two generated warnings into hard errors for a reason unrelated to
openSYDE. Doing it properly means moving each tool to target_compile_options --
eight CMakeLists files, gated on the full build, and worth landing on its own.

That matters because the -Wparentheses which would have caught the inverted
security-option comparisons was being emitted by these very builds, and read by
nobody.

Verified: core clean under -Werror, ctest 223/223, all eight tools build from
scratch (BUILD_RC=0) after rm -rf build result.

+++ opensyde_tool/src/system_views/system_update/update_package/C_SyvUpPacHexFileView.cpp
@@ -257,7 +257,7 @@ void C_SyvUpPacHexFileView::mh_AddApplicationInformation(C_OscHexFile & orc_HexF
    orc_Content += "</td>";
    orc_Content += "</tr>";
    orc_Content += "</table>";
-   for (int32_t s32_ItAppl = 0UL; s32_ItAppl < c_InfoBlocks.size(); ++s32_ItAppl)
+   for (int32_t s32_ItAppl = 0L; s32_ItAppl < static_cast<int32_t>(c_InfoBlocks.size()); ++s32_ItAppl)
    {
       const stw::opensyde_core::C_OscApplicationInfoBlock & rc_CurInfo = c_InfoBlocks[s32_ItAppl];
       orc_Content += "<h3>" +
